Separate the sync-related methods to allow explicitly passing in the sync method
